Add files via upload
Set up and configure a database for the future immnewsnetwork.com site that would accommodate for the features listed below.

a) Create the ability for public users to submit the contact form. Data should be stored in the database. When the form is submitted, the PHP script should display a "thank you" message.

b) Create a Content Management System for immnewsnetwork.com

- Create a login page for members.

One type of user is an administrator (At this time, manually add an administrator user record directly through PHPMyAdmin)
Create the ability for administrators to

1- Add, Edit and Delete articles
2 - Set the featured article
3 - Edit the text on the About page
4 - View Contact form submissions

c) Member functions

Allow public users to register as a memeber.Members will be able to:
- see a "like" button at the top of each article.All public users should see the total number of "likes" on each article.